refactor the package,methods and fix bugs
