+++ src/main/java/com/rocs/medical/records/application/app/facade/low/stock/medicine/LowStockMedicineFacade.java
@@ -1,0 +1,8 @@
+package com.rocs.medical.records.application.app.facade.low.stock.medicine;
+
+import com.rocs.medical.records.application.model.inventory.LowStockItem;
+
+import java.util.List;
+public interface LowStockMedicineFacade {
+    List<LowStockItem> findAllLowStockMedicine();
+}
