--- src/main/java/com/rocs/medical/records/application/app/facade/lowStockMedicine/LowStockMedicineFacade.java
@@ -1,8 +1,0 @@
-package com.rocs.medical.records.application.app.facade.lowStockMedicine;
-
-import com.rocs.medical.records.application.model.inventory.LowStockItem;
-
-import java.util.List;
-public interface LowStockMedicineFacade {
-    List<LowStockItem> checkLowStockAndNotify();
-}
