+++ src/main/java/com/rocs/medical/records/application/app/facade/reports/common/ailments/report/CommonAilmentsReportFacade.java
@@ -1,0 +1,29 @@
+package com.rocs.medical.records.application.app.facade.reports.common.ailments.report;
+
+import com.rocs.medical.records.application.model.reports.CommonAilmentsReport;
+
+import java.util.Date;
+import java.util.List;
+
+/**
+ * An interface that manages the CommonAilmentsReport object maintenance such as
+ * common ailments, occurrences, affected people, grade level, and strand
+ * within the given report period.
+ * */
+public interface CommonAilmentsReportFacade {
+
+    /**
+     * Generates a report listing common ailments within a given period
+     * @param startDate  The start date of the report period.
+     * @param endDate    The end date of the report period.
+     * @param gradeLevel The grade level to filter the report and can be null.
+     * @param section    The section to filter the report and can be null.
+     * @return A list of CommonAilmentsReport objects.
+     */
+    List<CommonAilmentsReport> generateCommonAilmentReport(
+            Date startDate,
+            Date endDate,
+            String gradeLevel,
+            String section
+    );
+}
