--- src/main/java/com/rocs/medical/records/application/data/dao/lowStockMedicine/LowStockMedicineDao.java
@@ -1,9 +1,0 @@
-package com.rocs.medical.records.application.data.dao.lowStockMedicine;
-
-import com.rocs.medical.records.application.model.inventory.LowStockItem;
-
-import java.util.List;
-
-public interface LowStockMedicineDao {
-    List<LowStockItem> getLowStockItems();
-}
